requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
gunicorn==23.0.0
supabase==2.15.0
python-dotenv==1.2.2
pydantic==2.12.5
httpx==0.28.1
anyio==4.13.0
APScheduler==3.10.4
